move componentes to make them reusable as well as some hooks
