+++ src/app/(protected)/analysis/[symbol]/components/NewsCarousel/NewsCarousel.tsx
@@ -4,7 +4,7 @@ import { stockNewsQuery } from '@/queries/stockNewsQuery';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import AutoSizer from 'react-virtualized-auto-sizer';
 import { FixedSizeList as List } from 'react-window';
-import { InfoCard } from '../InfoCard';
+import { InfoCard } from '../../../../../../components/InfoCard';
 import EmptyState from './components/EmptyState';
 import NewsCard from './components/NewsCard';
