+++ src/app/(protected)/analysis/[symbol]/components/Skeletons/BusinessInfo.tsx
@@ -1,6 +1,6 @@
 import { Separator } from '@/components/ui/separator';
 import { Skeleton } from '@/components/ui/skeleton';
-import { InfoCard } from '../InfoCard';
+import { InfoCard } from '../../../../../../components/InfoCard';
 
 export function BusinessInfoSkeleton() {
   return (
