+++ src/app/(protected)/analysis/[symbol]/components/Skeletons/Recommendations.tsx
@@ -1,5 +1,5 @@
 import { Skeleton } from '@/components/ui/skeleton';
-import { InfoCard } from '../InfoCard';
+import { InfoCard } from '../../../../../../components/InfoCard';
 
 export function RecommendationsSkeleton() {
   return (
