+++ src/components/EconomicIndicator.tsx
