+++ src/components/InfoCard.tsx
